+++ fractal/components/_preview-layouts/_stickies-preview.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+const page = ({ rendered_content, path }) => {
+
+    return (
+        <div>
+            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/core-configuration.css`}/>
+            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/stickies/StickyBar/StickyBar.css`}/>
+            <style>{"\
+                body{\
+                  height: 3000px;\
+                }\
+                .c-sticky{\
+                  position: relative;\
+                  top: 100px;\
+                  background-color: blue;\
+                }\
+              "}</style>
+            <div className="base-preview" dangerouslySetInnerHTML={{__html: rendered_content}}/>
+        </div>
+    );
+};
+
+module.exports = page;